Auto update state of SimplePager based Pagination
The Pagination widget now auto-updates it's state of the active button,
and the enabled state of the previous and next buttons.
This based on the state of the SimplePager.

Now it is now longer needed to call Pagination.rebuild manually each
time after the DataGrid or CellTable updates.

Also this reduces flickering as this does not clear and redraw the
complete Pagination widget like the rebuild method does.