Hide SuggestBox popup when suggestbox is no longer showing.
When SuggestBox is for example added to a Modal the modified method is
executed when the modal closes, and the suggestbox popup is open.

Now the Popup closes. In previous versions the Popup would stay open,
giving a weird effect. Or would even crash due to trying to set negative
css sizes.